src/microservices/events/requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
pydantic==2.6.1
kafka-python==2.0.2
python-dotenv==1.0.1